% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandsatDownload.R
\name{download_landsat_raster}
\alias{download_landsat_raster}
\title{Download Landsat Image as Raster}
\usage{
download_landsat_raster(start_date, end_date, roi, scale = 30)
}
\arguments{
\item{start_date}{Start date (e.g. "2023-01-01")}

\item{end_date}{End date (e.g. "2023-12-31")}

\item{roi}{Numeric vector of \link{minLon, minLat, maxLon, maxLat}}

\item{scale}{Resolution in meters (default is 30 for Landsat)}
}
\value{
A SpatRaster object
}
\description{
This function downloads a median composite of Landsat images over a date range and region.
}
