% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterDownload.R
\name{getSentinelData}
\alias{getSentinelData}
\title{Get Sentinel-2 Data from GEE}
\usage{
getSentinelData(start_date, end_date, roi, scale = 10)
}
\arguments{
\item{start_date}{Start date as string (YYYY-MM-DD)}

\item{end_date}{End date as string (YYYY-MM-DD)}

\item{roi}{Numeric vector of \link{minLon, minLat, maxLon, maxLat}}

\item{scale}{Scale/resolution in meters 10m for sentinel}

\item{via}{Method to fetch image:  "drive" ( safe export)}
}
\value{
A sentinel 2 image
}
\description{
Get Sentinel-2 Data from GEE
}
